Telegrambot/requirements.txt
python-dotenv
telethon
openai
rapidfuzz
flask
aiofiles
tenacity
snowballstemmer==3.0.1